native/all/pty/terminator_terminal_PtyProcess.cpp: src/terminator/terminal/PtyProcess.java: src/terminator/view/JTerminalPane.java: we can now implement PtyProcess.destroy without additional JNI. Revert the whimsical use of 'long' instead of 'int' for pid/getPid to avoid introducing casts.

+++ terminator/trunk/src/terminator/view/JTerminalPane.java
@@ -662,7 +662,7 @@ public class JTerminalPane extends JPanel {
 			return true;
 		}
 
-		final long directChildPid = ptyProcess.getPid();
+		final int directChildPid = ptyProcess.getPid();
 		final String processesUsingTty = ptyProcess.listProcessesUsingTty();
 
 		if (processesUsingTty.length() == 0) {
